% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCluster.R
\name{BioCluster}
\alias{BioCluster}
\title{Estimate the biological meaningfulness of cluster splits}
\usage{
BioCluster(
  obj,
  external,
  prefix = "RNA_snn_res.",
  logfc.threshold = 0.1,
  min.pct = 0.1,
  adjp = 0.01,
  topn = 20,
  bulk = TRUE
)
}
\arguments{
\item{obj}{a Seurat object with clustering at multiple resolutions}

\item{external}{a external dataset; bulk or one Seurat object for single cell RNA-seq; if bulk, row is the gene, column is the sample.}

\item{prefix}{the prefix of Seurat clustering resolution names (default: RNA_snn_res.)}

\item{logfc.threshold}{the cutoff of log2fold change to identify markers (default: 0.1)}

\item{min.pct}{the percent difference to identify markers (default: 0.1)}

\item{adjp}{the cutoff of adjusted p-value to identify markers (default: 0.01)}

\item{topn}{the number of marker genes (default: 20)}

\item{bulk}{indicate the external is from bulk RNAseq (TRUE) or single-cell RNA-seq (FALSE) (default:TRUE)}
}
\value{
a Seurat object with the score and p-value associated with each subcluster stored in misc$BioCluster
}
\description{
Estimate the support of one cluster grouped into several subclusters by the external dataset
}
