% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCluster.R
\name{PlotBioCluster}
\alias{PlotBioCluster}
\title{the BioCluster result Visualization}
\usage{
PlotBioCluster(
  obj,
  node.size = NULL,
  node.color = NULL,
  node.text = NULL,
  return = "plot"
)
}
\arguments{
\item{obj}{Seurat object}

\item{node.size}{node size; either a numeric value or NULL ;  if NULL, the size of each node is determined by the size of the cluster (default:NULL)}

\item{node.color}{node color; either a color or  NULL ; if NULL, the color of each node is determined by the resolution (default:NULL)}

\item{node.text}{node text; either "" or NULL; if NULL, the text is the cluster ID; if "", node text is not shown(default:NULL)}

\item{return}{specifying what to return, either "plot" (a ggplot object) or "graph" (a tbl_graph object) (default:"plot")}
}
\value{
a ggplot object or a tbl_graph object
}
\description{
Visualizing how clusters at a low resolution are grouped into several subclusters labeled with scores and p-values, suggesting the support of the partition by the external dataset.
}
